requirements.txt
joblib==1.5.0
matplotlib==3.10.3
pandas==2.2.3
Pillow==11.2.1
plotly==6.0.1
python-dotenv==1.1.0
scikit_learn==1.6.1
scipy==1.15.3
shap==0.47.2
streamlit==1.45.1
